fix: compile errors introduced by the restructuring atomization pass
The grep-based verification used throughout the restructuring waves (no
Android SDK/Kotlin compiler available in this environment) caught duplicate
declarations and dangling references, but missed several missing-import and
one type-mismatch bug that only a real compiler surfaces. Root-caused via
CI build logs and fixed:

- BiolismBackup/SessionHistory/TimerState.kt: missing DataStore edit()/remove()
  imports (needed the androidx.datastore.preferences.core.* wildcard, not
  just Preferences/DataStore).
- ScanOffLookup/ScanServerClient/RecipeServerImportClient.kt: ImagePayload
  imported from the wrong package (domain.model instead of
  data.remote.api).
- ActivityScreen.kt + ActivityDailyTotalsCard/AddActivityDialog.kt: missing
  LazyColumn/items and Modifier imports.
- TrackerKetosisFastingLogic/TrackerTimerControl/ResultSaveDestinations.kt:
  missing androidx.lifecycle.viewModelScope import (extension functions on a
  ViewModel still need it explicitly).
- MultiMarkerMonthGrid.kt: missing fillMaxSize import.
- DashboardGapLogging.kt: missing domain.engine.dashboard.* import (GapSuggestion).
- DiaryProteinPerSlotCard/ScanStateOverlay.kt: missing dp import.
- ActiveFastCard.kt: missing RoundedCornerShape import.
- GroceryItemRow.kt: missing cross-package CheckableGroceryItem import.
- GroceryProgressRow.kt: missing clip import.
- ProfileMetricsPreviewCard.kt: bmi parameter typed Int? instead of Double?
  (bmiValue is a Double? StateFlow), matching the pre-split behavior.
- ResultScanLoader.kt: missing computeMetabolics extension-function import.

Also updated scripts/check_scoring_drift.py: the atomized AdditivesDb.kt
(now assembled from per-tier files via `TIER1 + TIER2 + ...`) and OffMapper.kt
(several functions relocated to sibling files) broke the script's single-file
declaration extraction. Added a directory-wide fallback search for relocated
declarations, and a bare "IDENT + IDENT + ..." concatenation resolver that
splices the referenced vals' entries back together for comparison, so the
check still verifies the real data instead of failing on the assembly
expression's own shape.

+++ scan-eat-android/app/src/main/java/fr/scanneat/presentation/profile/components/ProfileMetricsPreviewCard.kt
@@ -27,7 +27,7 @@ import kotlin.math.roundToInt
 @Composable
 internal fun ProfileMetricsPreviewCard(
     currentProfile: Profile,
-    bmi: Int?,
+    bmi: Double?,
     bmiCat: fr.scanneat.domain.engine.scoring.BmiCategory?,
     tdee: Double?,
     tdeeGoal: Double?,
